Default to hiding filtered statuses
Tusky v21 and before didn't support the warn/hide distinction in filters, so everything was hidden.

Tusky v22.0 does, but since the default is "warn" for filters this means that they are now seeing warnings for content they've filtered, and the warnings include the very text they're trying to avoid.

Fix this by adding a new account-level preference that overrides the filter action for all their filters. Values are either NONE (use the filter's action), WARN (ignore the filter's action, show a warning), and HIDE (ignore the filter's action, hide the status).

The default value for the preference is HIDE, to match the v21.0 behaviour.

Fixes #3746