Handle placeholder items in a timeline list
This should not happen, as the lists do not use placeholders.

Nevertheless, in use I've observed `getItem(position)` returning null.

The documentation says this can happen if the item at that position is a placeholder.

Without this change this manifests as a completely empty status, but with buttons for voting, an empty media preview, etc.

With this change it's an effectively invisible view which is eventually replaced with the real one.