Composite semi-transparent avatars over a solid background
Avatars that are semi-transparent are a problem when viewing a thread, as the line that connects different statuses in the same thread is drawn underneath the avatar and is visible.

Fix this with a CompositeWithOpaqueBackground Glide transformation that:

1. Extracts the alpha channel from the avatar image
2. Converts the alpha to a 1bpp mask
3. Draws that mask on a new bitmap, with the appropriate background colour
4. Draws the original bitmap on top of that

So any partially transparent areas of the original image are drawn over a solid background colour, so anything drawn under them will not appear.