Display notification filter/clear actions as menu items
Previously the notification filter and clear actions were shown as buttons in the UI, with a preference that determined whether they were displayed.

Remove this preference, and display them as menu items.

- "Filter" is shown as an icon, if possible
- "Clear" is only ever shown as a menu item

If the user has disabled the top navigation bar a new options menu item is shown to the right of the tabs, whether they are at the top or bottom of the screen.