Show the correct tab title when opening the activity
The previous code used `notificationTabPosition`, which was never changed, so always 0.

This meant that if you e.g., got to `MainActivity` by clicking on a notification the title would show "Home".

Fix that by using the existing `position` variable which represents the currently selected tab, show the correct title is shown.

Fixes #3864.