Use setSupportActionBar more judiciously
Previous code either used setSupportActionBar (if the mainNav was enabled), or reproduced the logo/options menu functionality of an action bar if it was disabled.

There's no need to do that -- just have three toolbars (1 or 2 of which might be hidden), and use precisely one of them as the action bar.

Then you get logo support, navigation, and options menu handling for free.

This also simplifies the layouts and means you don't need to different bits of code for loading an image in to the action bar logo.