build.gradle: Always increase versionCode for Green
Due to Android (and the F-Droid) upgrade process, `versionCode` should always increase in order to trigger the update process. 

Use a dedicated `versionCode` for Green releases, time base, which will always increase (beware of time zones ⚠️ )

Another side effect will be: the same code, rebuilt, will be considered as an upgrade, as build time will be newer. 

It is impossible to use the commit number as it is not an `int`. 
A better idea would be to count the amounts of commits, with something like  `git rev-list --first-parent --count HEAD` but I didn't manage to do it.